add database entries for new gists, update REPL URLs
