+++ site/package.json
@@ -15,6 +15,7 @@
     "testsrc": "mocha -r esm test/**"
   },
   "dependencies": {
+    "@polka/redirect": "^1.0.0-next.0",
     "@polka/send": "^1.0.0-next.2",
     "devalue": "^1.1.0",
     "do-not-zip": "^1.0.0",
