+++ site/src/routes/repl/create.json.js
@@ -1,5 +1,5 @@
 import send from '@polka/send';
-import { body } from './_utils.js';
+import body from './_utils/body.js';
 import { query } from '../../utils/db';
 import { isUser } from '../../backend/auth';
